UI polish, ADS-B Ultrafeeder & stale cleanup
Frontend: visual polish and layout changes (floating header/env bar, tab overlays, tactical audio redesign, sidebar/header/ENV tweaks), dark map style, map controls moved and attribution disabled, aircraft rendering moved from a GeoJSON layer to DOM markers with hover labels and marker CSS, updated global styles (glassmorphism, marker visuals), vite dev server & proxy updated for local compose (host:0.0.0.0, port 5173, stream/op25 proxies), and docker-compose configured to run frontend in dev mode with mounted volume. .env.example: added ULTRAFEEDER_URL, ODOT_API_KEY, AIRNOW_API_KEY, NWS_ALERT_ZONES and guidance for ADS‑B source options. Websockets/store: added periodic purgeStaleEntities and a store action to remove stale aircraft (older than 60s). Poller/backend: publish_entity now sets TTL and includes improved error logging; AdsbPoller supports an ultrafeeder URL (prefers ULTRAFEEDER_URL or host/port) and _poll_ultrafeeder accepts a URL; db, geofence and weather normalizer changes add safer numeric casting and unit/format conversions. These changes enable optional local Ultrafeeder integration, reduce stale ADS‑B artifacts, and refresh the UI for a tactical dark theme.

+++ poller/geofence.py
@@ -24,7 +24,7 @@ async def check_geofences(entity: dict, conn) -> None:
         SELECT id, name, zone_type
         FROM geofences
         WHERE active = TRUE
-          AND ST_Contains(geom, ST_SetSRID(ST_MakePoint($1, $2), 4326))
+          AND ST_Contains(geom, ST_SetSRID(ST_MakePoint($1::float, $2::float), 4326))
         """,
         lon,
         lat,
